backend-server/ai/requirements.txt
# EchoMemory Azure AI Pipeline Requirements
azure-cognitiveservices-speech==1.33.0  # For STT & TTS
azure-ai-textanalytics==5.3.0          # For Entity Extraction & Sentiment
openai==1.3.5                          # For Azure OpenAI Narrative
mysql-connector-python==8.2.0
python-dotenv==1.0.0
